--- Lab2/README.md
@@ -1,9 +1,0 @@
-# Прикладное программирование
-
-## Лабораторная № 2. Обработка данных
-
-__Задание:__  
-__Вар 1-2:__ Реализовать классы и итераторы для получения данных из сформированного в ходе 1й лабораторной csv-файла.  
-__Вар 3-10:__ Реализовать классы и итераторы для формирования аннотации для полученных в ходе 1й лабораторной наборов изображений.
-
-Ноутбук с заданием и примерами кода доступен по [ссылке](https://github.com/itsecd/Application-Programming/blob/main/%D0%9B%D0%B0%D0%B1%D0%BE%D1%80%D0%B0%D1%82%D0%BE%D1%80%D0%BD%D0%B0%D1%8F%20%D1%80%D0%B0%D0%B1%D0%BE%D1%82%D0%B0%202.ipynb).
