--- Lab3/requirements.md
@@ -1,3 +1,0 @@
-PyQt6==6.6.1
-PyQt6-Qt6==6.6.1
-PyQt6-sip==13.6.0
